% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chart.line.R
\name{chart.line}
\alias{chart.line}
\title{Create Simple Line Charts of Georgian Stock Trades}
\usage{
chart.line(ID = "AGVQ", export = FALSE, type = "o", col = "red",
  time = "Date", y = "Min. price")
}
\arguments{
\item{ID}{character value of the single stock ID that we want trade results of}

\item{export}{boolean value of whether to export the data to a csv file}

\item{type}{take character string to decide what plot to draw. See ?plot for details on type}

\item{col}{take character string to decide what color to use. See ?plot for details on col}

\item{time}{take character string of either "Date" or "Session" to decide
whether to use date or session as x-axis}

\item{y}{determines what is the y-axis value;
       take character string of "Min. price" or "Max. price" or "Weighted avg. price" or "Weighted
       closing price" or "Volume in Securities" or "Volume in GEL"}
}
\description{
\code{chart.line} shows a specified line chart of a specific stock
}
\examples{
chart.line(ID = "AGVQ", export = FALSE, type = "l", col = "red", time = "Date", y = "Min. price")
}
\seealso{
\link{download.single.stock.trade}
}

