% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download.multi.stock.trade.R
\name{download.multi.stock.trade}
\alias{download.multi.stock.trade}
\title{Download Multiple Stocks Data from Georgian Stock Exchange Website}
\usage{
download.multi.stock.trade(ID)
}
\arguments{
\item{ID}{the data frame of stock IDs that we want trade results of.
           The data frame should contain only one column of symbols}
}
\value{
data frame containing the trade results of different stock issues from every day
        available on the website.
}
\description{
\code{download.multi.stock.trade} takes a data frame of stock IDs and returns a data frame of its trade result
}
\examples{
stocks = data.frame(ID = c("AGVQ", "AKAT"))
download.multi.stock.trade(ID = stocks)
}
\seealso{
\link{download.stockIDs}
}

