% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download.single.stock.trade.R
\name{download.single.stock.trade}
\alias{download.single.stock.trade}
\title{Download Single Stock Data from Georgian Stock Exchange Website}
\usage{
download.single.stock.trade(ID)
}
\arguments{
\item{ID}{the single stock ID that we want trade results of}
}
\value{
data frame containing the trade results of different stock issues from every day
        available on the website.
}
\description{
\code{download.single.stock.trade} takes a stock ID and returns a data frame of its trade result
}
\examples{
download.single.stock.trade(ID = "AGVQ")
}
\seealso{
\link{download.stockIDs}
}

