% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download.stockIDs.R
\name{download.stockIDs}
\alias{download.stockIDs}
\title{Download Stock IDs from Georgian Stock Exchange Website}
\usage{
download.stockIDs(export = FALSE)
}
\arguments{
\item{export}{boolean value of whether to export the data to a csv file}
}
\value{
data frame containing the stock IDs and corresponding company names
}
\description{
\code{download.stockIDs} returns a data frame of stock IDs and corresponding company names
}
\examples{
download.stockIDs()
}
\seealso{
\link{download.single.stock.trade}
}

