% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/export.csv.R
\name{export.csv}
\alias{export.csv}
\title{Export Stock Data to CSV Files}
\usage{
export.csv(data)
}
\arguments{
\item{data}{the data file to write into csv files}
}
\description{
\code{export.csv} writes a csv file for the stock data
}
\examples{
export.csv()
}
\seealso{
\link{download.all.stock.trade}
}

