% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download.all.stock.trade.R
\name{download.all.stock.trade}
\alias{download.all.stock.trade}
\title{Download All Stocks Data from Georgian Stock Exchange Website}
\usage{
download.all.stock.trade(export = FALSE)
}
\arguments{
\item{export}{boolean value of whether to export the data to a csv file}
}
\value{
data frame containing the trade results of all stocks issues from every day
        available on the website.
}
\description{
\code{download.all.stock.trade} returns a data frame of all stocks trade result
}
\details{
Notice that the first Session of all trade is 10, not 1. There is a Session 1
         according to GSE website, but it's blank
}
\examples{
\dontrun{
   download.all.stock.trade()
}
}
\seealso{
\link{download.stockIDs}
}

